Make it possible to set custom computer color
I was having problems distinguishing between hosts (e.g. between local and
remote hosts, or even between different remote hosts).

This commits adds customizability to the color the hostname appears in the
shell. There's a set of color variables in the script, and through the
`COMPUTER_COLOR` variable, you can customize the color of the hostname to any
of them.